+++ system/src/main/java/io/github/sds100/keymapper/system/media/AndroidMediaAdapter.kt
@@ -1,11 +1,13 @@
 package io.github.sds100.keymapper.system.media
 
+import android.content.ComponentName
 import android.content.Context
 import android.media.AudioAttributes
 import android.media.AudioManager
 import android.media.AudioPlaybackConfiguration
 import android.media.MediaPlayer
 import android.media.session.MediaController
+import android.media.session.MediaSessionManager
 import android.media.session.PlaybackState
 import androidx.core.content.getSystemService
 import androidx.core.net.toUri
@@ -13,6 +15,7 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import io.github.sds100.keymapper.common.utils.KMError
 import io.github.sds100.keymapper.common.utils.KMResult
 import io.github.sds100.keymapper.common.utils.Success
+import io.github.sds100.keymapper.system.notifications.NotificationReceiver
 import io.github.sds100.keymapper.system.volume.VolumeStream
 import java.io.FileNotFoundException
 import javax.inject.Inject
@@ -27,7 +30,7 @@ import kotlinx.coroutines.launch
 
 @Singleton
 class AndroidMediaAdapter @Inject constructor(
-    @ApplicationContext private val context: Context,
+    @ApplicationContext private val ctx: Context,
     private val coroutineScope: CoroutineScope,
 ) : MediaAdapter {
     companion object {
@@ -37,9 +40,15 @@ class AndroidMediaAdapter @Inject constructor(
         private const val SEEK_AMOUNT = 30000L
     }
 
-    private val ctx = context.applicationContext
-
+    private val mediaSessionManager: MediaSessionManager by lazy { ctx.getSystemService()!! }
     private val audioManager: AudioManager by lazy { ctx.getSystemService()!! }
+
+    private val notificationListenerComponent by lazy {
+        ComponentName(
+            ctx,
+            NotificationReceiver::class.java,
+        )
+    }
 
     private val activeMediaSessions: MutableStateFlow<List<MediaController>> =
         MutableStateFlow(emptyList())
@@ -72,121 +81,118 @@ class AndroidMediaAdapter @Inject constructor(
     }
 
     override fun fastForward(packageName: String?): KMResult<*> {
-        val controls = getPackageTransportControls(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (controls == null) {
-            return KMError.NoMediaSessions
-        } else {
-            controls.fastForward()
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_FAST_FORWARD)) {
+            session.transportControls.fastForward()
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
     override fun rewind(packageName: String?): KMResult<*> {
-        val controls = getPackageTransportControls(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (controls == null) {
-            return KMError.NoMediaSessions
-        } else {
-            controls.rewind()
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_REWIND)) {
+            session.transportControls.rewind()
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
     override fun play(packageName: String?): KMResult<*> {
-        val controls = getPackageTransportControls(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (controls == null) {
-            return KMError.NoMediaSessions
-        } else {
-            controls.play()
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_PLAY)) {
+            session.transportControls.play()
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
     override fun pause(packageName: String?): KMResult<*> {
-        val controls = getPackageTransportControls(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (controls == null) {
-            return KMError.NoMediaSessions
-        } else {
-            controls.pause()
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_PAUSE)) {
+            session.transportControls.pause()
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
     override fun playPause(packageName: String?): KMResult<*> {
-        val session = getPackageMediaSession(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (session == null) {
-            return KMError.NoMediaSessions
-        } else {
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_PLAY_PAUSE)) {
             when (session.playbackState?.state) {
                 PlaybackState.STATE_PLAYING -> session.transportControls.pause()
                 PlaybackState.STATE_PAUSED -> session.transportControls.play()
                 else -> {}
             }
-
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
     override fun previousTrack(packageName: String?): KMResult<*> {
-        val controls = getPackageTransportControls(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (controls == null) {
-            return KMError.NoMediaSessions
-        } else {
-            controls.skipToPrevious()
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_SKIP_TO_PREVIOUS)) {
+            session.transportControls.skipToPrevious()
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
     override fun nextTrack(packageName: String?): KMResult<*> {
-        val controls = getPackageTransportControls(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (controls == null) {
-            return KMError.NoMediaSessions
-        } else {
-            controls.skipToNext()
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_SKIP_TO_NEXT)) {
+            session.transportControls.skipToNext()
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
     override fun stop(packageName: String?): KMResult<*> {
-        val controls = getPackageTransportControls(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (controls == null) {
-            return KMError.NoMediaSessions
-        } else {
-            controls.stop()
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_STOP)) {
+            session.transportControls.stop()
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
     override fun stepForward(packageName: String?): KMResult<*> {
-        val session = getPackageMediaSession(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (session == null) {
-            return KMError.NoMediaSessions
-        } else {
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_SEEK_TO)) {
             val position = session.playbackState?.position ?: return KMError.NoMediaSessions
             session.transportControls.seekTo(position + SEEK_AMOUNT)
-
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
     override fun stepBackward(packageName: String?): KMResult<*> {
-        val session = getPackageMediaSession(packageName)
+        val session = getPackageMediaSession(packageName) ?: return KMError.NoMediaSessions
 
-        if (session == null) {
-            return KMError.NoMediaSessions
-        } else {
+        if (session.isPlaybackActionSupported(PlaybackState.ACTION_SEEK_TO)) {
             val position = session.playbackState?.position ?: return KMError.NoMediaSessions
-            val newPosition = max(position - SEEK_AMOUNT, 0)
-            session.transportControls.seekTo(newPosition)
+            session.transportControls.seekTo(max(0, position - SEEK_AMOUNT))
             return Success(Unit)
+        } else {
+            return KMError.MediaActionUnsupported
         }
     }
 
@@ -271,11 +277,16 @@ class AndroidMediaAdapter @Inject constructor(
         activeMediaSessions.update { mediaSessions }
     }
 
+    private fun MediaController.isPlaybackActionSupported(action: Long): Boolean =
+        (playbackState?.actions ?: 0) and action != 0L
+
     private fun getPackageMediaSession(packageName: String?): MediaController? {
+        val mediaSessions = mediaSessionManager.getActiveSessions(notificationListenerComponent)
+
         if (packageName == null) {
-            return activeMediaSessions.value.firstOrNull()
+            return mediaSessions.firstOrNull()
         } else {
-            for (session in activeMediaSessions.value) {
+            for (session in mediaSessions) {
                 if (session.packageName == packageName) {
                     return session
                 }
@@ -283,11 +294,5 @@ class AndroidMediaAdapter @Inject constructor(
         }
 
         return null
-    }
-
-    private fun getPackageTransportControls(
-        packageName: String?,
-    ): MediaController.TransportControls? {
-        return getPackageMediaSession(packageName)?.transportControls
     }
 }